refactor: useCardRegisterStates 훅 분리
